advanced_investment_crew/src/advanced_investment_crew/tools/technical_analysis_tool.py
from typing import Type

from crewai.tools import BaseTool
from pydantic import BaseModel, Field


class TechnicalAnalysisInput(BaseModel):
    symbol: str = Field(..., description="Ticker symbol")
    time_horizon: str = Field(default="1y", description="Time horizon for analysis")


class TechnicalAnalysisTool(BaseTool):
    name: str = "Technical Analysis Tool"
    description: str = "Performs technical analysis on a given symbol."
    args_schema: Type[BaseModel] = TechnicalAnalysisInput

    def _run(self, symbol: str, time_horizon: str = "1y") -> dict:
        # Placeholder outputs; replace with actual TA computations (MA, RSI, MACD, etc.)
        return {
            "symbol": symbol,
            "time_horizon": time_horizon,
            "trend": "pending_data",
            "key_levels": [],
            "indicators": {},
        }




