advanced_investment_crew/pyproject.toml
[project]
name = "advanced_investment_crew"
version = "0.1.0"
description = "Advanced investment analysis crew project"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.28.8",
    "crewai-tools>=0.3.0",
    "yfinance>=0.2.50",
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "scikit-learn>=1.4.2",
    "langchain-openai>=0.2.2",
    "chromadb>=0.5.4",
    "langchain-community>=0.3.0",
    "PyYAML>=6.0.1",
    "streamlit>=1.34.0",
    "forex-python>=1.8",
    "pycountry>=22.3.0",
    "plotly>=5.20.0",
    "scipy>=1.10.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
run_crew = "advanced_investment_crew.main:run"
train_crew = "advanced_investment_crew.main:train"
replay_crew = "advanced_investment_crew.main:replay"
test_crew = "advanced_investment_crew.main:test"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
advanced_investment_crew = ["config/*.yaml"]

